service/requirements.txt
alembic==1.5.8
aniso8601==9.0.1
astroid==2.4.2
attrs==20.3.0
bcrypt==3.2.0
certifi==2020.11.8
cffi==1.14.5
chardet==3.0.4
click==7.1.2
colorama==0.4.4
cryptography==3.4.7
Flask==1.1.2
Flask-API==2.0
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.9
Flask-Migrate==2.7.0
Flask-RESTful==0.3.8
flask-restx==0.3.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.5.1
Flask-Testing==0.8.1
greenlet==1.0.0
idna==2.10
isort==5.6.4
itsdangerous==1.1.0
Jinja2==2.11.3
joblib==1.0.1
jsonschema==3.2.0
lazy-object-proxy==1.4.3
Mako==1.1.4
MarkupSafe==1.1.1
mccabe==0.6.1
numpy==1.21.0
pandas==1.2.5
psycopg2-binary==2.8.6
pycparser==2.20
PyJWT==2.0.1
pylint==2.6.0
pyrsistent==0.17.3
python-dateutil==2.8.1
python-dotenv==0.17.0
python-editor==1.0.4
pytz==2021.1
requests==2.25.0
scikit-learn==0.24.2
scipy==1.7.0
six==1.15.0
sklearn==0.0
SQLAlchemy==1.4.5
threadpoolctl==2.1.0
toml==0.10.2
urllib3==1.26.2
Werkzeug==1.0.1
wrapt==1.12.1
